Rename the registry related files.
