+++ isaaclab_arena/environments/isaaclab_interop.py
@@ -5,7 +5,7 @@
 
 import argparse
 
-from isaaclab_arena.assets.asset_registry import EnvironmentRegistry
+from isaaclab_arena.assets.registries import EnvironmentRegistry
 from isaaclab_arena_environments.cli import ensure_environments_registered
