+++ isaaclab_arena/examples/compile_env_notebook.py
@@ -16,7 +16,7 @@ simulation_app = AppLauncher()
 
 # %%
 
-from isaaclab_arena.assets.asset_registry import AssetRegistry
+from isaaclab_arena.assets.registries import AssetRegistry
 from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
 from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
 from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
