+++ isaaclab_arena/tests/test_contact_sensor_not_at_root.py
@@ -19,8 +19,8 @@ def get_test_environment(num_envs: int):
     # This test tests that we can successfully at a contact sensor to an object
     # whose rigid body is not at the root of the USD file.
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
     from isaaclab_arena.assets.object_reference import ObjectReference
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.embodiments.franka.franka import FrankaIKEmbodiment
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
