+++ isaaclab_arena/tests/test_detect_object_type.py
@@ -62,8 +62,8 @@ def _test_detect_object_type(simulation_app):
 
 
 def _test_detect_object_type_for_all_objects(simulation_app):
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
     from isaaclab_arena.assets.object_utils import detect_object_type
+    from isaaclab_arena.assets.registries import AssetRegistry
 
     asset_registry = AssetRegistry()
     for object_asset in asset_registry.get_assets_by_tag("object"):
@@ -87,9 +87,9 @@ def _test_detect_object_type_for_all_objects(simulation_app):
 
 def _test_auto_object_type(simulation_app):
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
     from isaaclab_arena.assets.object import Object
     from isaaclab_arena.assets.object_base import ObjectType
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
     from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
