+++ isaaclab_arena/tests/test_dexsuite_kuka_lift_example.py
@@ -13,7 +13,7 @@ def test_dexsuite_lift_example_in_cli_registry() -> None:
     pytest.importorskip(
         "isaaclab_tasks.manager_based.manipulation.dexsuite.config.kuka_allegro.dexsuite_kuka_allegro_env_cfg"
     )
-    from isaaclab_arena.assets.asset_registry import EnvironmentRegistry
+    from isaaclab_arena.assets.registries import EnvironmentRegistry
     from isaaclab_arena_environments.cli import ensure_environments_registered
 
     ensure_environments_registered()
@@ -27,7 +27,7 @@ def test_procedural_assets_registered() -> None:
     pytest.importorskip(
         "isaaclab_tasks.manager_based.manipulation.dexsuite.config.kuka_allegro.dexsuite_kuka_allegro_env_cfg"
     )
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
+    from isaaclab_arena.assets.registries import AssetRegistry
 
     reg = AssetRegistry()
     assert reg.is_registered("procedural_table")
@@ -40,7 +40,7 @@ def test_dexsuite_kuka_lift_task_matches_lift_mdp_flags() -> None:
         "isaaclab_tasks.manager_based.manipulation.dexsuite.config.kuka_allegro.dexsuite_kuka_allegro_env_cfg"
     )
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.metrics.success_rate import SuccessRateMetric
     from isaaclab_arena.tasks.lift_object_task import DexsuiteLiftTask, DexsuiteLiftTerminationsCfg, LiftObjectTask
     from isaaclab_arena.utils.pose import Pose, PoseRange
