+++ isaaclab_arena/tests/test_events.py
@@ -24,8 +24,8 @@ def _test_set_object_pose_per_env_event(simulation_app):
 
     from isaaclab.managers import EventTermCfg, SceneEntityCfg
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
     from isaaclab_arena.assets.object_reference import ObjectReference
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
     from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
@@ -118,7 +118,7 @@ def _test_object_moves_with_initial_velocity(simulation_app):
     """Test that a sphere moves with the given initial velocity after reset."""
     import isaaclab.sim as sim_utils
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
     from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
