+++ isaaclab_arena/tests/test_galbot_embodiment.py
@@ -18,9 +18,9 @@ INITIAL_POSITION_EPS = 1e-6
 
 def get_galbot_test_environment(num_envs: int = 1):
     """Returns a kitchen scene with Galbot embodiment for testing."""
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
     from isaaclab_arena.assets.object_base import ObjectType
     from isaaclab_arena.assets.object_reference import ObjectReference
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.embodiments.common.arm_mode import ArmMode
     from isaaclab_arena.embodiments.galbot.galbot import GalbotEmbodiment
