+++ isaaclab_arena/tests/test_locomanip_object_on_termination.py
@@ -16,7 +16,7 @@ def _test_g1_locomanip_object_on_destination_termination(simulation_app) -> bool
 
     from isaaclab.managers import SceneEntityCfg
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
     from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
