+++ isaaclab_arena/tests/test_object_on_termination.py
@@ -21,8 +21,8 @@ def _test_object_on_destination_termination(simulation_app) -> bool:
 
     from isaaclab.managers import SceneEntityCfg
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
     from isaaclab_arena.assets.object_reference import ObjectReference
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.embodiments.franka.franka import FrankaIKEmbodiment
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
