+++ isaaclab_arena/tests/test_object_pose_randomization.py
@@ -17,7 +17,7 @@ HEADLESS = True
 def _test_object_pose_randomization(simulation_app):
     """Test that object poses are randomized within the specified range."""
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
     from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
