+++ isaaclab_arena/tests/test_revolute_joint_moved_rate_metric.py
@@ -27,7 +27,7 @@ EXPECTED_MOVEMENT_RATE_EPS = 1e-6
 def _test_revolute_joint_moved_rate(simulation_app):
     """Returns a scene which we use for these tests."""
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
     from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
