+++ isaaclab_arena/tests/test_robot_initial_position.py
@@ -20,7 +20,7 @@ INITIAL_POSITION_EPS = 1e-6
 def _test_robot_initial_position(simulation_app):
     """Returns a scene which we use for these tests."""
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
     from isaaclab_arena.environments.isaaclab_arena_environment import IsaacLabArenaEnvironment
