+++ isaaclab_arena/tests/test_sequential_open_door.py
@@ -16,7 +16,7 @@ HEADLESS = True
 def get_test_environment(remove_reset_door_state_event: bool, num_envs: int):
     """Returns a scene which we use for these tests."""
 
-    from isaaclab_arena.assets.asset_registry import AssetRegistry
+    from isaaclab_arena.assets.registries import AssetRegistry
     from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
     from isaaclab_arena.embodiments.franka.franka import FrankaIKEmbodiment
     from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
